move TestHelper to Uti
