+++ src/main/java/FixCache/Util/TestHelper.java
@@ -1,4 +1,4 @@
-package Test;
+package Util;
 
 import java.io.FileInputStream;
 import java.sql.Connection;
