+++ src/test/java/FixCache/Test/CacheReplacementTest.java
@@ -12,6 +12,7 @@ import org.junit.Test;
 import Cache.Cache;
 import Cache.CacheReplacement;
 import Cache.CacheItem.CacheReason;
+import Util.TestHelper;
 
 public class CacheReplacementTest {
